% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_axes.R
\name{add_axes}
\alias{add_axes}
\title{Add axes to plot}
\usage{
add_axes(a1, a2, x_dim, z_dim, y_dim, x_axis_col, z_axis_col, y_axis_col, data)
}
\arguments{
\item{a1}{projection angle a1}

\item{a2}{projection angle a2}

\item{x_dim}{Length of x dimension (image width)}

\item{z_dim}{Length of z dimension (image height)}

\item{y_dim}{Length of y dimension (image height out of plane)}

\item{x_axis_col}{x axis colour}

\item{z_axis_col}{z axis colour}

\item{y_axis_col}{y axis colour}

\item{data}{dataframe created during processing of image}
}
\description{
Intended for internal use only
}
