% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_height_plane.R
\name{add_height_plane}
\alias{add_height_plane}
\title{Add height plane to plot}
\usage{
add_height_plane(a1, a2, x_dim, z_dim, y_dim, col, lty, fill, alpha)
}
\arguments{
\item{a1}{Angle of rotation around the vertical axis in degrees}

\item{a2}{Angle of rotation around the horizontal axis in degrees}

\item{x_dim}{Length of x dimension (image width)}

\item{z_dim}{Length of z dimension (image height)}

\item{y_dim}{Maximal height of cuboids}

\item{col}{colour for border of height plane}

\item{lty}{linetype for border of height plane}

\item{fill}{fill colour for height plane}

\item{alpha}{transparency of height plane}
}
\description{
Intended for internal use only
}
