% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cuboid_image.R
\name{cuboid_image}
\alias{cuboid_image}
\title{Produce pseudo 3-D images from pixel values of 2-D images}
\usage{
cuboid_image(
  img,
  res = 60,
  resize_filter = "Mitchell",
  height_scale = NULL,
  height_map = v,
  cuboid_fill = NULL,
  cuboid_col = NA,
  orientation = 1,
  crop_square = TRUE,
  crop_gravity = "center",
  a1 = 45,
  a2 = asin(tan(pi/6)) * (180/pi),
  shading = c(0, 0.25, 0.5),
  return_data = FALSE,
  show_axes = TRUE,
  axis_cols = c("orange", "blue", "red"),
  show_height_plane = FALSE,
  height_plane_fill = NA,
  height_plane_col = axis_cols[3],
  height_plane_lty = "longdash",
  height_plane_alpha = 0.1,
  verbose = FALSE
)
}
\arguments{
\item{img}{a magick image or image file path/URL}

\item{res}{\code{img} is resized to be \code{res} pixels wide. A value of NULL keeps original size. (default = 60)}

\item{resize_filter}{resize filter to be used \code{magick::filter_types} (default = 'Mitchell')}

\item{height_scale}{a two value vector defining the lower and upper heights for scaling (default = c(1, 10))}

\item{height_map}{an expression containing r, g, b, h, s, v, x, or y (default = v)}

\item{cuboid_fill}{a vector of colours to be mapped to cuboid height (defaults to the original pixel colour)}

\item{cuboid_col}{a colour for the cuboid stroke outline}

\item{orientation}{orientation of the image prior to being projected (1, 2, 3 or 4) (default = 1)}

\item{crop_square}{crop image to a square (default = TRUE)}

\item{crop_gravity}{gravity of square crop if \code{crop_square = TRUE} (value from \code{magick::gravity_types()})}

\item{a1}{angle of rotation around the vertical axis in degrees (default = 45)}

\item{a2}{angle of rotation around the horizontal axis in degrees (default = asin(tan(pi/6))*(180/pi) ~35.264)}

\item{shading}{a vector of 3 numbers between 0 and 1. The amount of shading applied to each cuboid face. Order is top, left, right. (Default = c(0, 0.25, 0.5))}

\item{return_data}{return dataframe instead of image (default = FALSE)}

\item{show_axes}{show axes on the output image? (default = TRUE)}

\item{axis_cols}{a vector of colours for the three axes}

\item{show_height_plane}{show a polygon plane at the max height of the image (default = FALSE)}

\item{height_plane_fill}{colour for the height plane fill}

\item{height_plane_col}{colour for the height plane stroke outline}

\item{height_plane_lty}{line type for the height plane stroke outline}

\item{height_plane_alpha}{alpha value applied to the fill colour of the height plane}

\item{verbose}{print messages during processing}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Convert image pixels into cuboids that are raised from the ground plane and view as isometric projection (by default)
}
\details{
By default cuboids are raised in proportion to pixel brightness on a scale between 1 and 10.
Shading of the cuboid faces is also done naively - it's a simple reduction in brightness, so there is no clever physics going on here!
Some further reading on the Mitchell-Netravali resize filter https://www.cs.utexas.edu/~fussell/courses/cs384g-fall2013/lectures/mitchell/Mitchell.pdf and here https://legacy.imagemagick.org/Usage/filter/#mitchell
}
\examples{
cuboid_image(
    img = 'https://www.r-project.org/Rlogo.png',
    crop_square = FALSE,
    height_map = h)
}
