% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cuboid_matrix.R
\name{cuboid_matrix}
\alias{cuboid_matrix}
\title{Produce pseudo 3-D images from matrices}
\usage{
cuboid_matrix(
  mat,
  cuboid_fill = NULL,
  cuboid_col = NA,
  orientation = 1,
  a1 = 45,
  a2 = asin(tan(pi/6)) * (180/pi),
  shading = c(0, 0.25, 0.5),
  return_data = FALSE,
  show_axes = TRUE,
  axis_cols = c("orange", "blue", "red"),
  show_height_plane = FALSE,
  height_plane_fill = NA,
  height_plane_col = axis_cols[3],
  height_plane_lty = "longdash",
  height_plane_alpha = 0.1,
  verbose = FALSE
)
}
\arguments{
\item{mat}{a matrix}

\item{cuboid_fill}{a vector of colours to be mapped to cuboid height (for images this defaults to the original pixel colour)}

\item{cuboid_col}{a colour for the cuboid stroke outline}

\item{orientation}{orientation of the output (1, 2, 3 or 4) (default = 1)}

\item{a1}{angle of rotation around the vertical axis in degrees (default = 45)}

\item{a2}{angle of rotation around the horizontal axis in degrees (default = asin(tan(pi/6))*(180/pi) ~35.264)}

\item{shading}{a vector of 3 numbers between 0 and 1. The amount of shading applied to each cuboid face. Order is top, left, right. (Default = c(0, 0.25, 0.5))}

\item{return_data}{return dataframe instead of image (default = FALSE)}

\item{show_axes}{show axes on the output image? (default = TRUE)}

\item{axis_cols}{a vector of colours for the three axes}

\item{show_height_plane}{show a polygon plane at the max height of the image? (default = FALSE)}

\item{height_plane_fill}{colour for the height plane fill}

\item{height_plane_col}{colour for the height plane stroke outline}

\item{height_plane_lty}{line type for the height plane stroke outline}

\item{height_plane_alpha}{alpha value applied to the fill colour of the height plane}

\item{verbose}{print messages during processing}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Convert a numeric matrix into cuboids that are raised from the ground plane and view as isometric projection (by default)
}
\examples{
cuboid_matrix(matrix(runif(100), nrow = 10), cuboid_col = 1)

}
