% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isocuboids-package.R
\docType{package}
\name{isocuboids-package}
\alias{isocuboids}
\alias{isocuboids-package}
\title{isocuboids: Production Of Pseudo 3-D Isometric Height Maps}
\description{
A package for transforming images/pictures or matrices into pseudo 3D isometric height maps.
}
\author{
\strong{Maintainer}: Chris Holmes \email{c.holmes4@gmail.com}

}
\keyword{internal}
