% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso-project.R
\name{project}
\alias{project}
\title{Compute projected coordinates}
\usage{
project(x, y, z, va, ha)
}
\arguments{
\item{x}{x coordinate}

\item{y}{y coordinate}

\item{z}{z coordinate}

\item{va}{angle of rotation around the vertical axis (degrees)}

\item{ha}{angle of rotation arounf the horizontal axis (degrees)}
}
\description{
Intended for internal use only
https://en.wikipedia.org/wiki/Isometric_projection
}
