% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{unit_cube_df}
\alias{unit_cube_df}
\title{The unprojected unit cube dataframe}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 12 rows and 4 columns.
}
\usage{
unit_cube_df
}
\description{
The building block of all isocuboid images
}
\keyword{datasets}
